Recalibrate kinect and add inverted celebrate_0

+++ Arduino/Flew/msg.ino
@@ -12,7 +12,7 @@ int parseCoords(int n, int data[][2]) {
 }
 
 void onNum(void) {
-  int n = c.readBinArg<int>();
+  long n = c.readBinArg<int>();
   animate(n);
 }
